profileInspector + refactor
