+++ src/features/index.ts
@@ -4,6 +4,7 @@ import memberUpdateLogger from './memberUpdateLogger'
 import nameChecker from './nameChecker'
 import ping from './ping'
 import preventEmojiSpam from './preventEmojiSpam'
+import profileInspector from './profileInspector'
 import runtimeConfig from './runtimeConfig'
 import stickyMessage from './stickyMessage'
 import threadPruner from './threadPruner'
@@ -13,6 +14,7 @@ export default [
   nameChecker,
   ping,
   preventEmojiSpam,
+  profileInspector,
   runtimeConfig,
   stickyMessage,
   threadPruner,
